% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearListeqls.R, R/simpeCap.R
\name{clr}
\alias{clr}
\alias{clr}
\title{Function clr()}
\usage{
clr()

clr()
}
\description{
Function to rm all objects in global environment

Function to rm all objects in global environment
}
\examples{
clr()
clr()
}
\keyword{list}
\keyword{ls}
\keyword{rm}
