\name{is_rackable}
\alias{is_rackable}
\title{
Test for Rackable applications
}
\description{
A convenience function for testing whether or not objects are either a function or reference class as defined by the Rack specification for applicaions.
}
\usage{
is_rackable(app)
}
\arguments{
  \item{app}{ Any R object.}
}
\value{
    Logical determining whether or not argument is Rackable. Not vectorized.
}
\keyword{function}
