% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallpackages.R
\name{listpackages}
\alias{listpackages}
\title{Function listpackages()}
\usage{
listpackages(version = 3.4)
}
\arguments{
\item{version}{Defaults to 3.4.}
}
\description{
Function to get the list of installed packages (previous R version)
}
\examples{
listpackages()
}
\keyword{list}
\keyword{packages}
