\name{Multipart-class}
\Rdversion{1.1}
\docType{class}
\alias{Multipart-class}
\alias{Multipart}

\title{Class \code{"Multipart"}}
\description{
A convenience object for parsing multipart/form-data POST payloads.
}
\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("Multipart")
}
\keyword{classes}
\section{Methods}{
  \describe{
    \item{\code{parse(env)}:}{ Returns parsed POST payload as a named list.}
  }
}
